Add support for features in feature sets and feats.
